% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor.R
\name{json_path}
\alias{json_path}
\title{Apply JSONPath to a json object}
\usage{
json_path(json, path, strict = TRUE, zero_index = TRUE)
}
\arguments{
\item{json}{a JSON object loaded
by \code{\link[jsonlite]{read_json}}}

\item{path}{a JSONPath expression string,
see details}
}
\value{
a \code{data.frame} of JSONPath results
}
\description{
Takes a JSON object as loaded via the
\code{read_json} function and applies
a JSONPath query.
}
\details{
JSONPath is an XPath-like language for
querying JSON data.
}
