% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{read_json}
\alias{read_json}
\title{Re-exported read_json function}
\usage{
read_json(path, simplifyVector = FALSE, ...)
}
\description{
This function is imported from the
\code{jsonlite} package, see its documentation
there: \code{\link[jsonlite]{read_json}}.
}
