% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjsonpath.R
\docType{package}
\name{rjsonpath-package}
\alias{rjsonpath}
\alias{rjsonpath-package}
\title{rjsonpath: a JSONPath implementation for R}
\description{
rjsonpath: a JSONPath implementation for R
}
\author{
\strong{Maintainer}: Ben Moore \email{ben@blm.io}

}
