% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor.R
\name{json_path}
\alias{json_path}
\title{Apply JSONPath to a json object}
\usage{
json_path(json, path, strict = TRUE, zero_index = TRUE, simplify = TRUE)
}
\arguments{
\item{json}{a JSON object loaded
by \code{\link[jsonlite]{read_json}}}

\item{path}{a JSONPath expression string,
see details}

\item{strict}{require JSONPath starts with '$.'}

\item{zero_index}{numeric indices start from 0
rather than 1}

\item{simplify}{convert simple list results to
vectors}
}
\value{
a list or vector of JSONPath results
}
\description{
Takes a JSON object as loaded via the
\code{read_json} function and applies
a JSONPath query.
}
\details{
JSONPath is an XPath-like language for
querying JSON data, see \url{http://goessner.net/articles/JsonPath/}
for details.
}
\examples{
\dontrun{
json <- read_json("bookstore.json")
authors <- json_path(json, "$.store.books[*].author")
# [1] "Nigel Rees"       "Evelyn Waugh"     "Herman Melville"  "J. R. R. Tolkien"
}


}
